algo-trader/requirements.txt
kiteconnect
pandas
numpy
requests
beautifulsoup4
transformers
torch
ta-lib
python-dotenv
streamlit
plotly
sqlalchemy
websocket-client
schedule
